move types.ts to store/types.ts
