+++ packages/webapp/src/game/drawInitialHand.ts
@@ -6,7 +6,7 @@
 
 import { Hash } from "src/chain"
 import { mimcHash } from "src/utils/hashing"
-import { PrivateInfo } from "src/types"
+import { PrivateInfo } from "src/store/types"
 import { INITIAL_HAND_SIZE, MAX_DECK_SIZE, MAX_HAND_SIZE } from "src/game/constants"
 import { bigintToHexString, parseBigInt } from "src/utils/js-utils"
