+++ packages/webapp/src/store/network.ts
@@ -10,7 +10,7 @@ import { readContract } from "wagmi/actions"
 import { gameABI, inventoryABI } from "src/generated"
 import { deployment } from "src/deployment"
 import { type Fetched, throttledFetch } from "src/utils/throttled-fetch"
-import { FetchedGameData } from "src/types"
+import { FetchedGameData } from "src/store/types"
 import { Address } from "src/chain"
 
 // =================================================================================================
