+++ packages/webapp/src/store/read.ts
@@ -10,8 +10,8 @@
 
 import { type Address } from "src/chain"
 import * as store from "src/store/atoms"
-import type { FetchedGameData, PlayerData, PrivateInfo } from "src/types"
-import { GameStatus } from "src/types"
+import type { FetchedGameData, PlayerData, PrivateInfo } from "src/store/types"
+import { GameStatus } from "src/store/types"
 
 // =================================================================================================
 // BASIC STORE ACCESS
